+++ lib/teiserver/communication/libs/room_message_lib.ex
@@ -71,7 +71,7 @@ defmodule Teiserver.Communication.RoomMessageLib do
           sender_id: sender_id,
           room_id: room_id,
           content: content,
-          inserted_at: Timex.now()
+          inserted_at: DateTime.utc_now()
         },
         attrs
       )