+++ test/communication/libs/room_message_lib_test.exs
@@ -10,7 +10,7 @@ defmodule Teiserver.RoomMessageLibTest do
   defp valid_attrs do
     %{
       content: "some content",
-      inserted_at: Timex.now(),
+      inserted_at: DateTime.utc_now(),
       sender_id: AccountFixtures.user_fixture().id,
       room_id: CommunicationFixtures.room_fixture().id
     }
@@ -19,7 +19,7 @@ defmodule Teiserver.RoomMessageLibTest do
   defp update_attrs do
     %{
       content: "some updated content",
-      inserted_at: Timex.now(),
+      inserted_at: DateTime.utc_now(),
       sender_id: AccountFixtures.user_fixture().id,
       room_id: CommunicationFixtures.room_fixture().id
     }