+++ test/settings/server_setting_queries_test.exs
@@ -33,10 +33,10 @@ defmodule Teiserver.ServerSettingQueriesTest do
             key: "key1",
             value: ["value1", "value2"],
             value: "value1",
-            inserted_after: Timex.now(),
-            inserted_before: Timex.now(),
-            updated_after: Timex.now(),
-            updated_before: Timex.now()
+            inserted_after: DateTime.utc_now(),
+            inserted_before: DateTime.utc_now(),
+            updated_after: DateTime.utc_now(),
+            updated_before: DateTime.utc_now()
           ],
           order_by: [
             "Newest first",