+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added concept of user choices to represent pre-game choices made by users (e.g. in-game faction)
 - Removed clustering code so it can be handled by the application using Teiserver as a library
 - Moved everything from `Teiserver.Api` to `Teiserver`
+- Dropped Timex
 
 ## v0.0.4
 - Added `Lobby`, `LobbySummary`, `MatchType`, `Match`, `MatchMembership`, `MatchSettingType`, `MatchSetting` schemas, libs and queries